fix: Torch MPS backend failing test (#20709)

+++ keras/src/utils/backend_utils_test.py
@@ -15,7 +15,7 @@ class BackendUtilsTest(testing.TestCase):
     )
     def test_dynamic_backend(self, name):
         dynamic_backend = backend_utils.DynamicBackend()
-        x = np.random.uniform(size=[1, 2, 3])
+        x = np.random.uniform(size=[1, 2, 3]).astype("float32")
 
         if name == "numpy":
             dynamic_backend.set_backend(name)
